+++ src/standards.ts
@@ -1,161 +1,113 @@
-import util from 'util'
-import { write, eventEmitter } from './write'
+import fs from 'fs'
+import Unlink from './unlink'
+import RmDir from './rmdir'
 
-// Emit event with better message error
-function eventError (err: NodeJS.ErrnoException, file: string): void {
-  if (err) {
-    switch (err.code) {
-      case 'EMFILE':
-        eventEmitter.emit('warn', file, `Too many open files, cannot ${err.syscall || 'access'}: `)
-        break
-      case 'ENOENT':
-        eventEmitter.emit('warn', file, 'This file no longer exists: ')
-        break
-      case 'EPERM':
-        eventEmitter.emit('error', file, `Operation not permitted on this file (${err.syscall}): `)
-        break
-      default:
-        if (err.message === '64bit files are not yet supported.') {
-          eventEmitter.emit('error', file, '64bit files are not yet supported.')
-        } else {
-          console.log(err)
-          throw err
-        }
-      // break
-    }
-  }
-}
-
-type Callback = (err: NodeJS.ErrnoException | null) => void
-
-interface ArgsUnlinkStandard {
+interface StandardArgs {
   name?: string
   passes?: number
   description?: string
-  method: (file: string, callback: Callback) => void
+  unlinkStandard: typeof fs.unlink
+  rmdirStandard?: typeof fs.rmdir
 }
 
-class UnlinkStandard {
-  readonly name: string
-  readonly passes: number
-  readonly description: string
-  // FIXME method type is any
-  readonly method: any
-  constructor ({ name = 'Standard #', passes = 1, description = 'no description', method }: ArgsUnlinkStandard) {
-    this.name = name
-    this.passes = passes
-    this.description = description
-    this.method = method
-    this.method.__promisify__ = function (file: string) {
-      return util.promisify(this.method)
-    }
+export class Standard {
+  name: string
+  passes: number
+  description: string
+  unlinkStandard: typeof fs.unlink
+  rmdirStandard: typeof fs.rmdir
+
+  constructor ({ name, passes, description, unlinkStandard, rmdirStandard }: StandardArgs) {
+    this.name = name || 'Standard #'
+    this.passes = passes || 1
+    this.description = description || 'no description'
+    this.unlinkStandard = unlinkStandard
+    this.rmdirStandard = rmdirStandard || fs.rmdir
   }
 }
 
 // Object listing every standards
-const standards = {
-  randomData: new UnlinkStandard({
+export const standards = {
+  log: new Standard({
+    name: 'Log',
+    passes: 1,
+    description: 'Display targeted files.',
+    unlinkStandard: new Unlink()
+      .log()
+      .compile,
+    rmdirStandard: new RmDir()
+      .log()
+      .compile
+  }),
+
+  randomData: new Standard({
     name: 'Pseudorandom data',
     passes: 1,
     description: `Also kwown as "Australian Information Security Manual Standard ISM 6.2.92"
 and "New Zealand Information and Communications Technology Standard NZSIT 402" 
 Your data is overwritten with cryptographically strong pseudo-random data. (The data is indistinguishable from random noise.)`,
-    method: function (file: string, callback: Callback) {
-      write.init(file)
-        .then(({ file, fileSize }) => write.random(file, fileSize))
-        .then(({ file }) => write.unlink(file))
-        .then(() => callback(null))
-        .catch((err: NodeJS.ErrnoException) => {
-          eventError(err, file)
-          callback(err)
-        })
-    }
-
+    unlinkStandard: new Unlink()
+      .random()
+      .unlink()
   }),
-  randomByte: new UnlinkStandard({
+
+  randomByte: new Standard({
     name: 'Pseudorandom byte',
     passes: 1,
     description: 'Overwriting with a random byte.',
-    method: function (file: string, callback: Callback) {
-      write.init(file)
-        .then(({ file, fileSize }) => write.randomByte(file, fileSize))
-        .then(({ file }) => write.unlink(file))
-        .then(() => callback(null))
-        .catch((err: NodeJS.ErrnoException) => {
-          eventError(err, file)
-          callback(err)
-        })
-    }
+    unlinkStandard: new Unlink()
+      .randomByte()
+      .unlink()
   }),
-  zeroes: new UnlinkStandard({
+
+  zeroes: new Standard({
     name: 'Zeroes',
     passes: 1,
     description: 'Overwriting with zeroes.',
-    method: function (file: string, callback: Callback) {
-      write.init(file)
-        .then(({ file, fileSize }) => write.zeroes(file, fileSize))
-        .then(({ file }) => write.unlink(file))
-        .then(() => callback(null))
-        .catch((err: NodeJS.ErrnoException) => {
-          eventError(err, file)
-          callback(err)
-        })
-    }
+    unlinkStandard: new Unlink()
+      .zeroes()
+      .unlink()
   }),
-  ones: new UnlinkStandard({
+
+  ones: new Standard({
     name: 'Ones',
     passes: 1,
     description: 'Overwriting with ones.',
-    method: function (file: string, callback: Callback) {
-      write.init(file)
-        .then(({ file, fileSize }) => write.ones(file, fileSize))
-        .then(({ file }) => write.unlink(file))
-        .then(() => callback(null))
-        .catch((err: NodeJS.ErrnoException) => {
-          eventError(err, file)
-          callback(err)
-        })
-    }
+    unlinkStandard: new Unlink()
+      .ones()
+      .unlink()
   }),
-  secure: new UnlinkStandard({
+
+  secure: new Standard({
     name: '**Secure-rm standard**',
     passes: 3,
     description:
       `Pass 1: Overwriting with random data;
 Pass 2: Renaming the file with random data;
 Pass 3: Truncating between 25% and 75% of the file.`,
-    method: function (file: string, callback: Callback) {
-      write.init(file)
-        .then(({ file, fileSize }) => write.random(file, fileSize))
-        .then(({ file, fileSize }) => write.rename(file, fileSize))
-        .then(({ file, fileSize }) => write.truncate(file, fileSize))
-        .then(({ file }) => write.unlink(file))
-        .then(() => callback(null))
-        .catch((err: NodeJS.ErrnoException) => {
-          eventError(err, file)
-          callback(err)
-        })
-    }
+    unlinkStandard: new Unlink()
+      .random()
+      .rename()
+      .truncate()
+      .unlink(),
+    rmdirStandard: new RmDir()
+      .rename()
+      .rmdir()
   }),
-  'GOST_R50739-95': new UnlinkStandard({
+
+  'GOST_R50739-95': new Standard({
     name: 'Russian State Standard GOST R50739-95',
     passes: 2,
     description:
       `Pass 1: Overwriting with zeroes;
 Pass 2: Overwriting with random data.`,
-    method: function (file: string, callback: Callback) {
-      write.init(file)
-        .then(({ file, fileSize }) => write.zeroes(file, fileSize))
-        .then(({ file, fileSize }) => write.random(file, fileSize))
-        .then(({ file }) => write.unlink(file))
-        .then(() => callback(null))
-        .catch((err: NodeJS.ErrnoException) => {
-          eventError(err, file)
-          callback(err)
-        })
-    }
+    unlinkStandard: new Unlink()
+      .zeroes()
+      .random()
+      .unlink()
   }),
-  HMG_IS5: new UnlinkStandard({
+
+  HMG_IS5: new Standard({
     name: 'British HMG Infosec Standard 5',
     passes: 3,
     description:
@@ -166,40 +118,28 @@ and "Navy Staff Office Publication NAVSO P-5239-26"
 Pass 1: Overwriting with zeroes;
 Pass 2: Overwriting with ones;
 Pass 3: Overwriting with random data as well as verifying the writing of this data.`,
-    method: function (file: string, callback: Callback) {
-      write.init(file)
-        .then(({ file, fileSize }) => write.zeroes(file, fileSize))
-        .then(({ file, fileSize }) => write.ones(file, fileSize))
-        .then(({ file, fileSize }) => write.random(file, fileSize))
-        .then(({ file }) => write.unlink(file))
-        .then(() => callback(null))
-        .catch((err: NodeJS.ErrnoException) => {
-          eventError(err, file)
-          callback(err)
-        })
-    }
+    unlinkStandard: new Unlink()
+      .zeroes()
+      .ones()
+      .random()
+      .unlink()
   }),
-  'AR380-19': new UnlinkStandard({
+
+  'AR380-19': new Standard({
     name: 'US Army AR380-19',
     passes: 3,
     description:
       `Pass 1: Overwriting with random data;
 Pass 2: Overwriting with a random byte;
 Pass 3: Overwriting with the complement of the 2nd pass, and verifying the writing.`,
-    method: function (file: string, callback: Callback) {
-      write.init(file)
-        .then(({ file, fileSize }) => write.random(file, fileSize))
-        .then(({ file, fileSize }) => write.randomByte(file, fileSize))
-        .then(({ file, fileSize }) => write.complementary(file, fileSize))
-        .then(({ file }) => write.unlink(file))
-        .then(() => callback(null))
-        .catch((err: NodeJS.ErrnoException) => {
-          eventError(err, file)
-          callback(err)
-        })
-    }
+    unlinkStandard: new Unlink()
+      .random()
+      .randomByte()
+      .complementary()
+      .unlink()
   }),
-  VSITR: new UnlinkStandard({
+
+  VSITR: new Standard({
     name: 'Standard of the Federal Office for Information Security (BSI-VSITR)',
     passes: 7,
     description:
@@ -208,60 +148,42 @@ Pass 1: Overwriting with zeroes;
 Pass 2: Overwriting with ones;
 Pass 3-6: Same as 1-2;
 Pass 7: Overwriting with a random data as well as review the writing of this character.`,
-    method: function (file: string, callback: Callback) {
-      write.init(file)
-        .then(({ file, fileSize }) => write.zeroes(file, fileSize))
-        .then(({ file, fileSize }) => write.ones(file, fileSize))
-        .then(({ file, fileSize }) => write.zeroes(file, fileSize))
-        .then(({ file, fileSize }) => write.ones(file, fileSize))
-        .then(({ file, fileSize }) => write.zeroes(file, fileSize))
-        .then(({ file, fileSize }) => write.ones(file, fileSize))
-        .then(({ file, fileSize }) => write.random(file, fileSize))
-        .then(({ file }) => write.unlink(file))
-        .then(() => callback(null))
-        .catch((err: NodeJS.ErrnoException) => {
-          eventError(err, file)
-          callback(err)
-        })
-    }
+    unlinkStandard: new Unlink()
+      .zeroes()
+      .ones()
+      .zeroes()
+      .ones()
+      .zeroes()
+      .ones()
+      .random()
+      .unlink()
   }),
-  schneier: new UnlinkStandard({
+
+  schneier: new Standard({
     name: 'Bruce Schneier Algorithm',
     passes: 7,
     description:
       `Pass 1: Overwriting with zeros;
 Pass 2: Overwriting with ones;
 Pass 3-7: Overwriting with random data.`,
-    method: function (file: string, callback: Callback) {
-      write.init(file)
-        .then(({ file, fileSize }) => write.zeroes(file, fileSize))
-        .then(({ file, fileSize }) => write.ones(file, fileSize))
-        .then(({ file, fileSize }) => write.random(file, fileSize, 5))
-        .then(({ file }) => write.unlink(file))
-        .then(() => callback(null))
-        .catch((err: NodeJS.ErrnoException) => {
-          eventError(err, file)
-          callback(err)
-        })
-    }
+    unlinkStandard: new Unlink()
+      .zeroes()
+      .ones()
+      .random(5)
+      .unlink()
   }),
-  pfitzner: new UnlinkStandard({
+
+  pfitzner: new Standard({
     name: 'Pfitzner Method',
     passes: 33,
     description:
       'Pass 1-33: Overwriting with random data.',
-    method: function (file: string, callback: Callback) {
-      write.init(file)
-        .then(({ file, fileSize }) => write.random(file, fileSize, 33))
-        .then(({ file }) => write.unlink(file))
-        .then(() => callback(null))
-        .catch((err: NodeJS.ErrnoException) => {
-          eventError(err, file)
-          callback(err)
-        })
-    }
+    unlinkStandard: new Unlink()
+      .random(33)
+      .unlink()
   }),
-  gutmann: new UnlinkStandard({
+
+  gutmann: new Standard({
     name: 'Peter Gutmann Algorithm',
     passes: 35,
     description:
@@ -273,43 +195,24 @@ Pass 10-25: Overwriting with 0x00, incremented by 1 at each pass, until 0xFF;
 Pass 26-28: Same as 7-9;
 Pass 29-31: Overwriting with 0x6D 0xB6 0xDB, then cycling through the bytes;
 Pass 32-35: Overwriting with random data.`,
-    method: function (file: string, callback: Callback) {
-      write.init(file)
-        .then(({ file, fileSize }) => write.random(file, fileSize, 4))
-        .then(({ file, fileSize }) => write.byte(file, 0x55, fileSize))
-        .then(({ file, fileSize }) => write.byte(file, 0xAA, fileSize))
-        .then(({ file, fileSize }) => write.cycleBytes(file, [0x92, 0x49, 0x24], fileSize))
-        .then(({ file, fileSize }) => write.incrementByte(file, { start: 0x00, end: 0xFF, increment: 0x11 }, fileSize))
-        .then(({ file, fileSize }) => write.cycleBytes(file, [0x92, 0x49, 0x24], fileSize))
-        .then(({ file, fileSize }) => write.cycleBytes(file, [0x6D, 0xB6, 0xDB], fileSize))
-        .then(({ file, fileSize }) => write.random(file, fileSize, 4))
-        .then(({ file }) => write.unlink(file))
-        .then(() => callback(null))
-        .catch((err: NodeJS.ErrnoException) => {
-          eventError(err, file)
-          callback(err)
-        })
-    }
-  })/*
-  T: new UnlinkStandard({
-    name: 'Template',
-    passes: 3,
-    description:
-      `description`,
-    method: function (file: string, callback: Callback) {
-      write.init(file)
-        .then((fileSize) => write.random(file, fileSize))
-        .then(() => write.unlink(file))
-        .then(() => callback(null))
-        .catch((err: NodeJS.ErrnoException) => {
-          logError(err, file)
-          callback(err)
-        })
-    }
-  }), */
+    unlinkStandard: new Unlink()
+      .random(4)
+      .byte(0x55)
+      .byte(0xAA)
+      .byteArray([0x92, 0x49, 0x24])
+      .byteArray([0x49, 0x24, 0x92])
+      .byteArray([0x24, 0x92, 0x49])
+      .forByte({ init: 0x00, condition: i => i < 0xFF, increment: i => i + 0x11 })
+      .byteArray([0x92, 0x49, 0x24])
+      .byteArray([0x49, 0x24, 0x92])
+      .byteArray([0x24, 0x92, 0x49])
+      .byteArray([0x6D, 0xB6, 0xDB])
+      .byteArray([0xB6, 0xDB, 0x6D])
+      .byteArray([0xDB, 0x6D, 0xB6])
+      .random(4)
+      .unlink()
+  })
 }
 
 // List of valid standards IDs
-const validIDs = <unknown>Object.keys(standards) as keyof typeof standards
-
-export { standards, eventEmitter, validIDs, UnlinkStandard }
+export const validIDs = <unknown>Object.keys(standards) as keyof typeof standards