+++ CHANGELOG.md
@@ -14,8 +14,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Custom directory removal.
+
 ### Changed
 - Repository location.
+- New custom standard syntax.
 
 ### Security
 - Libraries update.